requirements.txt
# project requirement
fastapi
uvicorn[standard]
google-generativeai
chromadb
python-dotenv
pydantic